+++ src/routes.js
@@ -6,7 +6,7 @@ import {
 } from 'react-navigation';
 
 import Login from './views/Login'
-import Dashboard from './views/Dashboard'
+import Dashboard from './views/Dashboard/Dashboard'
 
 const appNavigator = createStackNavigator({
     login : { screen : Login, navigationOptions:{header:null}},
